% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.download.R
\name{eco.download}
\alias{eco.download}
\title{Download clean Fred or BLS Economic data}
\usage{
eco.download(source, keys, shape, periodicity, show.recession, time.series)
}
\arguments{
\item{source}{Determines API to use: fred or bls. This must be defined.}

\item{keys}{bls or fred keys. Fred keys can be found on the Fred website. BLS keys have been cleaned to be human readable ie labor_force}

\item{shape}{Choose 'long' for long format or 'wide' for wide format. Default is wide.}

\item{periodicity}{Choose data periodicity. Either 'monthly', 'quarterly', 'weekly', or 'daily'. Default is monthly.}

\item{show.recession}{Choose y to include recession indicator or 'n' to exclude. Default is 'y'.}

\item{time.series}{Choose 'y' to convert dataframe into a time series (xts) object}
}
\value{
df Cleaned dataframe is returned
}
\description{
fredr and blsAPI wrapper to download and format economic data.
}
\examples{
\dontrun{eco.download(keys=c('UNRATE','PAYEMS'),  show.recession = 'y', time.series = 'y', periodicity='monthly')}
}
