% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.plot.R
\name{eco.plot}
\alias{eco.plot}
\title{Plot Economic Data}
\usage{
eco.plot(
  plot_type,
  df,
  x,
  y1,
  y2,
  y3,
  labels.list,
  theme.list,
  stat,
  bar.position,
  fill.value,
  ycolor1,
  ycolor2,
  ycolor3,
  date.break,
  date.format
)
}
\arguments{
\item{plot_type}{Type of plot to create. Options are line, dual line, tri line, bar, and bar line}

\item{df}{Dataframe containing economic data}

\item{x}{x axis}

\item{y1}{first y axis}

\item{y2}{Optional 2nd y axis}

\item{y3}{Optional 3rd y axis}

\item{labels.list}{List containing chart labs}

\item{stat}{Optional bar parameter}

\item{bar.position}{Optional bar parameter}

\item{fill.value}{Optional bar parameter}

\item{ycolor1}{Optional first y color}

\item{ycolor2}{Optional second y color}

\item{date.break}{# Optional date break ie '1 month'}

\item{date.format}{Optional date format ie '%y'}

\item{theme_list}{Optional List containing theme options}

\item{ycolor}{3 Optional third y color}
}
\value{
df dataframe with transformations applied
}
\description{
Quickly plot economic data with recession bars
}
\examples{
eco.plot(plot_type = 'line', df, x='date', y1= value.UNRATE)
}
