% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.transform.R
\name{eco.transform}
\alias{eco.transform}
\title{Transform Economic Data}
\usage{
eco.transform(df, transformation, lags)
}
\arguments{
\item{df}{Dataframe containing economic data that will be transformed.}

\item{transformation}{The type of transformation to be applied.}

\item{lags}{The number of lags to apply}
}
\value{
df dataframe with transformations applied
}
\description{
Quickly transform an entire economic dataset with a variety of transformations
}
\examples{
\dontrun{eco.transform(df = data, transformation = 'pct_chg', lags = 1)}
}
