% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.api.key.R
\name{fred.api.key}
\alias{fred.api.key}
\title{Fredr API Key Wrapper}
\usage{
fred.api.key(key)
}
\arguments{
\item{key}{Fred API Key}
}
\description{
Wrapper to set Fred API Key
}
\examples{
fred.api.key(##########)
}
