% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recession.probability.R
\name{recession.probability}
\alias{recession.probability}
\title{Recession Probability Based on Treasury Spread}
\usage{
recession.probability(start.date)
}
\arguments{
\item{start.date}{Date to begin analysis}
}
\value{
graph
}
\description{
Compute and visualize the most recent recession probability
}
\examples{
recession.probability('2000-01-01')
}
