% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.download.R
\name{eco.download}
\alias{eco.download}
\title{Download clean Fred or BLS Economic data}
\usage{
eco.download(source, keys, shape, periodicity, show.recession, time.series)
}
\arguments{
\item{source}{Determines API to use: fred or bls. This must be defined.}

\item{keys}{bls or fred keys. Fred keys can be found on the Fred website. BLS keys have been cleaned to be human readable ie labor_force}

\item{shape}{Choose 'long' for long format or 'wide' for wide format. Default is wide.}

\item{periodicity}{Choose data periodicity. Either 'monthly', 'quarterly', 'weekly', or 'daily'. Default is monthly.}

\item{show.recession}{Choose y to include recession indicator or 'n' to exclude. Default is 'y'.}

\item{time.series}{Choose 'y' to convert dataframe into a time series (xts) object}
}
\value{
df Cleaned dataframe is returned
}
\description{
fredr and blsAPI wrapper to download and format economic data.
}
\details{
## BLS Data Codes
labor_force
employment
unemployment_rate
nonfarm_employment
avg_weekly_hrs
avg_weekly_hrs_nonsup
nonfarm_productivity
nonfarm_labor_costs
nonfarm_hrly_compensation
nonfarm_multi_productivity
cpi_urban_80s
cpu_urban_67
core_cpi_u
core_cpi_w
ppi_final_sa
ppi_final_ua
ppi_final_core
ppi_final_core_less_trade
ppi_finished_goods_82
imports_all_commodities
exports_all_commodities
employment_cost_index
eci_private
eci_private_wages
}
\examples{
df <-eco.download(c('UNRATE','PAYEMS'),  show.recession = 'y', time.series = 'y', periodicity='monthly')
}
